requirements.txt
dateutils
python-dateutil
requests
tqdm
pyyaml
